Ghosts should be fine

+++ src/t8_default/t8_dprism_bits.c
@@ -171,18 +171,18 @@ t8_dprism_boundary_face (const t8_dprism_t * p, int face,
   }
   switch (face) {
   case 0:
-    q->x = p->tri.y * P4EST_ROOT_LEN / T8_DTRI_ROOT_LEN;
-    q->y = p->line.x *  P4EST_ROOT_LEN / T8_DLINE_ROOT_LEN;
+    q->x = ((int64_t)p->tri.y * P4EST_ROOT_LEN) / T8_DTRI_ROOT_LEN;
+    q->y = ((int64_t)p->line.x *  P4EST_ROOT_LEN) / T8_DLINE_ROOT_LEN;
     q->level = p->tri.level;
     break;
   case 1:
-    q->x = p->tri.x *   P4EST_ROOT_LEN / T8_DTRI_ROOT_LEN;
-    q->y = p->line.x * P4EST_ROOT_LEN / T8_DLINE_ROOT_LEN;
+    q->x = ((int64_t)p->tri.x *   P4EST_ROOT_LEN) / T8_DTRI_ROOT_LEN;
+    q->y = ((int64_t)p->line.x * P4EST_ROOT_LEN) / T8_DLINE_ROOT_LEN;
     q->level = p->tri.level;
     break;
   case 2:
-    q->x = p->tri.x *   P4EST_ROOT_LEN / T8_DTRI_ROOT_LEN;
-    q->y = p->line.x * P4EST_ROOT_LEN / T8_DLINE_ROOT_LEN;
+    q->x = ((int64_t)p->tri.x *   P4EST_ROOT_LEN) / T8_DTRI_ROOT_LEN;
+    q->y = ((int64_t)p->line.x * P4EST_ROOT_LEN) / T8_DLINE_ROOT_LEN;
     q->level = p->tri.level;
     break;
   default:
@@ -327,22 +327,22 @@ t8_dprism_extrude_face (const t8_element_t * face, t8_element_t * elem,
     p->line.level = q->level;
     p->tri.level = q->level;
     p->tri.x = T8_DTRI_ROOT_LEN - T8_DTRI_LEN (p->tri.level);
-    p->tri.y = q->x * T8_DTRI_ROOT_LEN / P4EST_ROOT_LEN;
-    p->line.x = q->y * T8_DLINE_ROOT_LEN / P4EST_ROOT_LEN;
+    p->tri.y = ((int64_t)q->x * T8_DTRI_ROOT_LEN) / P4EST_ROOT_LEN;
+    p->line.x = ((int64_t)q->y * T8_DLINE_ROOT_LEN) / P4EST_ROOT_LEN;
     break;
   case 1:
     p->line.level = q->level;
     p->tri.level = q->level;
-    p->tri.x = q->x * T8_DTRI_ROOT_LEN / P4EST_ROOT_LEN;
-    p->tri.y = q->x * T8_DTRI_ROOT_LEN / P4EST_ROOT_LEN;
-    p->line.x = q->y * T8_DLINE_ROOT_LEN / P4EST_ROOT_LEN;
+    p->tri.x = ((int64_t)q->x * T8_DTRI_ROOT_LEN) / P4EST_ROOT_LEN;
+    p->tri.y =((int64_t) q->x * T8_DTRI_ROOT_LEN )/ P4EST_ROOT_LEN;
+    p->line.x = ((int64_t)q->y * T8_DLINE_ROOT_LEN) / P4EST_ROOT_LEN;
     break;
   case 2:
     p->line.level = q->level;
     p->tri.level = q->level;
-    p->tri.x = q->x * T8_DTRI_ROOT_LEN / P4EST_ROOT_LEN;;
+    p->tri.x = ((int64_t)q->x * T8_DTRI_ROOT_LEN) / P4EST_ROOT_LEN;;
     p->tri.y = 0;
-    p->line.x = q->y * T8_DLINE_ROOT_LEN / P4EST_ROOT_LEN;
+    p->line.x =((int64_t) q->y * T8_DLINE_ROOT_LEN) / P4EST_ROOT_LEN;
     break;
   case 3:
     p->line.level = t->level;
